% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-filter-score.R
\name{as_score_vec}
\alias{as_score_vec}
\title{Coerce to a \code{score_vec} object}
\usage{
as_score_vec(x, direction = "maximize", impute = Inf)
}
\arguments{
\item{x}{A numeric vector..}

\item{direction}{What is the best result? Possible values are "maximize",
"minimize", or "zero".}

\item{impute}{A single numeric value for when values are missing. This should
be a value that indicates that a variable \emph{should not} be removed.}
}
\description{
\code{as_score_vec()} provides coercion to \code{score_vec} from other
existing objects.
}
\examples{
as_score_vec(1:4)
}
