% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_corr.R
\docType{data}
\name{filter_corr}
\alias{filter_corr}
\alias{filter_corr_rank}
\title{Correlation filter}
\format{
An object of class \code{filter_method_corr} (inherits from \code{filter_method}) of length 6.

An object of class \code{filter_method_corr_rank} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_corr

filter_corr_rank
}
\description{
Correlation filter
}
\keyword{datasets}
