% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pls.R
\docType{data}
\name{filter_imp_pls}
\alias{filter_imp_pls}
\title{Partial least squares importance filters}
\format{
An object of class \code{filter_method_imp_pls} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_imp_pls
}
\description{
Partial least squares importance filters
}
\keyword{datasets}
