% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_imp.R
\docType{data}
\name{filter_imp_rf}
\alias{filter_imp_rf}
\alias{filter_imp_crf}
\title{Random forests importance filters}
\format{
An object of class \code{filter_method_imp_rf} (inherits from \code{filter_method}) of length 6.

An object of class \code{filter_method_imp_crf} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_imp_rf

filter_imp_crf
}
\description{
Random forests importance filters
}
\keyword{datasets}
