% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_info.R
\docType{data}
\name{filter_info_gain}
\alias{filter_info_gain}
\alias{filter_info_gain_ratio}
\alias{filter_mic}
\title{Information theory filters}
\format{
An object of class \code{filter_method_info_gain} (inherits from \code{filter_method}) of length 6.

An object of class \code{filter_method_info_gain_ratio} (inherits from \code{filter_method}) of length 6.

An object of class \code{filter_method_mic} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_info_gain

filter_info_gain_ratio

filter_mic
}
\description{
Information theory filters
}
\keyword{datasets}
