% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_methods}
\alias{filter_methods}
\title{Find available importance filters}
\usage{
filter_methods()
}
\value{
A character vector of methods.
}
\description{
This function shows the methods that can be used with \code{\link[=importance_metrics]{importance_metrics()}}.
}
\details{
\itemize{
\item \code{corr}: The correlation between numeric predictors and a numeric outcome.
\item \code{corr_rank}: The Spearman (rank) correlation between numeric predictors and
a numeric outcome.
\item \code{imp_crf}: The unbiased variable importance score from a conditional
inference random forest model. Any type of predictor or outcome can be used.
\item \code{imp_pls}:
\item \code{imp_rf}: The variable importance score from a standard random
forest model. Any type of predictor or outcome can be used.
\item \code{info_gain}:
\item \code{info_gain_ratio}:
\item \code{max_diff}:
\item \code{mic}:
\item \code{mrmr}:
\item \code{roc_auc}: The area under the receiver operating characteristic (ROC) curve
for numeric predictors and a binary outcome.
}
}
\examples{
filter_methods()
}
