% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mrmr.R
\docType{data}
\name{filter_mrmr}
\alias{filter_mrmr}
\title{Minimum Redundancy Maximal Relevancy filter}
\format{
An object of class \code{filter_method_mrmr} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_mrmr
}
\description{
Minimum Redundancy Maximal Relevancy filter
}
\keyword{datasets}
