% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_roc_auc.R
\docType{data}
\name{filter_roc_auc}
\alias{filter_roc_auc}
\title{Area under the ROC curve filter}
\format{
An object of class \code{filter_method_roc_auc} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_roc_auc
}
\description{
Area under the ROC curve filter
}
\keyword{datasets}
