% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdb-package.R
\docType{package}
\name{filterdb-package}
\alias{filterdb}
\alias{filterdb-package}
\title{filterdb: Methods for Supervised Filtering of Predictors}
\description{
Low-level data and developer interfaces for supervised filtering methods.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software PBC [copyright holder]
}

}
\keyword{internal}
