% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_corr.R, R/filter_imp_pls.R,
%   R/filter_imp_rf.R, R/filter_info.R, R/filter_max_diff.R, R/filter_mrmr.R,
%   R/filter_roc_auc.R
\name{fit_xy.filter_method_corr}
\alias{fit_xy.filter_method_corr}
\alias{fit_xy.filter_method_corr_rank}
\alias{fit_xy.filter_method_imp_pls}
\alias{fit_xy.filter_method_imp_rf}
\alias{fit_xy.filter_method_imp_crf}
\alias{fit_xy.filter_method_info_gain}
\alias{fit_xy.filter_method_info_gain_ratio}
\alias{fit_xy.filter_method_mic}
\alias{fit_xy.filter_method_max_diff}
\alias{fit_xy.filter_method_mrmr}
\alias{fit_xy.filter_method_roc_auc}
\title{Execute a single supervised filter}
\usage{
\method{fit_xy}{filter_method_corr}(object, x, y, ...)

\method{fit_xy}{filter_method_corr_rank}(object, x, y, ...)

\method{fit_xy}{filter_method_imp_pls}(object, x, y, ...)

\method{fit_xy}{filter_method_imp_rf}(object, x, y, seed = sample.int(1000, 1), ...)

\method{fit_xy}{filter_method_imp_crf}(object, x, y, seed = sample.int(1000, 1), ...)

\method{fit_xy}{filter_method_info_gain}(object, x, y, ...)

\method{fit_xy}{filter_method_info_gain_ratio}(object, x, y, ...)

\method{fit_xy}{filter_method_mic}(object, x, y, ...)

\method{fit_xy}{filter_method_max_diff}(object, x, y, ...)

\method{fit_xy}{filter_method_mrmr}(object, x, y, ...)

\method{fit_xy}{filter_method_roc_auc}(object, x, y, ...)
}
\arguments{
\item{object}{An object of class \code{filter_method}.}

\item{x}{A data frame of predictors.}

\item{y}{A data frame with the outcome column.}

\item{...}{Options passed to the underlying method (if any).}

\item{seed}{A single integer to set the seed. If \code{NULL}, the current RNG
state is used.}
}
\value{
A tibble with a character column (\code{variable}) and a column for the
score values.
}
\description{
Execute a single supervised filter
}
\keyword{internal}
