% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_tools.R
\name{get_score_info}
\alias{get_score_info}
\alias{get_score_info.filter_method}
\alias{get_score_info.filter_results}
\title{Get some attribute of a filter method}
\usage{
get_score_info(x, att = "goal", ...)

\method{get_score_info}{filter_method}(x, att = "goal", ...)

\method{get_score_info}{filter_results}(x, att = "goal", ...)
}
\arguments{
\item{x}{An object.}

\item{att}{A single character value for the attribute being returned.}

\item{...}{Not currently used.}
}
\description{
Get some attribute of a filter method
}
