% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_metrics.R
\name{importance_metrics}
\alias{importance_metrics}
\title{Compute univariate scores of predictors}
\usage{
importance_metrics(x, y, methods, opts = list())
}
\arguments{
\item{x}{A data frame of predictors.}

\item{y}{A data frame with a single outcome column.}

\item{methods}{A character string of available scores. See \code{\link[=filter_methods]{filter_methods()}}.}

\item{A}{named list of options for the chosen \code{methods}.}
}
\value{
An object with class \code{importance_metrics}. The elements are \code{values}
(a list of data frames of scores) and \code{info} (a list with the details of each
method).
}
\description{
Compute univariate scores of predictors
}
\examples{
if (rlang::is_installed(c("dplyr", "modeldata"))) {
  library(dplyr)
  library(modeldata)

  ames_scores <-
    importance_metrics(
      ames \%>\% select(-Sale_Price),
      y = ames \%>\% select(Sale_Price),
      methods = c("corr")
    )

  ames_scores
}

}
