% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-filter-score.R
\name{is_score_vec}
\alias{is_score_vec}
\title{Test if an object inherits from \code{score_vec}}
\usage{
is_score_vec(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\code{is_score_vec()} checks if an object is a \code{score_vec} object.
}
\examples{

x <- score_vec(1:5)

is_score_vec(x)

}
