% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdb-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{obj_sum}
\alias{fit_xy}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit_xy}}, \code{\link[generics]{tidy}}}

  \item{pillar}{\code{\link[pillar:type_sum]{obj_sum}}}
}}

