% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-vec.R
\name{return-direction}
\alias{return-direction}
\alias{direction}
\alias{missing_val}
\alias{impute_score}
\title{Return desired direction}
\usage{
direction(x)

missing_val(x)

impute_score(x)
}
\arguments{
\item{x}{A \code{score_vec} object.}
}
\value{
A single character value (for the entire vector)
}
\description{
Return desired direction
}
\examples{
set.seed(1)
roc_values <- score_vec(runif(10, min = 1 / 2), "maximize")
direction(roc_values)

try(direction(1L))

}
