% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-vec.R
\name{score-helpers}
\alias{score-helpers}
\alias{direction}
\alias{missing_val}
\alias{impute_score}
\alias{transform.score_vec}
\title{Helpers for score vectors}
\usage{
direction(x)

missing_val(x)

impute_score(x)

\method{transform}{score_vec}(`_data`, ...)
}
\arguments{
\item{x, _data}{A \code{score_vec} object.}

\item{...}{Not used.}
}
\value{
\itemize{
\item \code{direction()} returns single character value (for the entire vector)
\item \code{missing_val()} returns a numeric value that corresponds to a score that
will prevent any predictor from being removed.
\item \code{impute_score()} returns a \code{score_vec} with any missing values filled in
with the results of \code{missing_val()}.
\item \code{transform()} returns a \code{score_vec} with the values transformed in line
it's goal. For goals of \code{"maximize_abs"}, \code{"minimize_abs"}, and \code{"zero"}
the absolute value of the score is returned.
}
}
\description{
Helpers for score vectors
}
\examples{
set.seed(1)
cor_values <- score_vec(runif(10, min = -1), direction = "maximize_abs", impute = 1.0)
direction(cor_values)

transform(cor_values)
missing_val(cor_values)

cor_values_miss <- c(cor_values, NA_real_)
impute_score(cor_values_miss)

}
