% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_metrics.R
\name{tidy.importance_metrics}
\alias{tidy.importance_metrics}
\title{Tidy importance metrics}
\usage{
\method{tidy}{importance_metrics}(x, impute = TRUE, transform = TRUE, ...)
}
\arguments{
\item{x}{An object produced by \code{\link[=importance_metrics]{importance_metrics()}}.}

\item{impute}{Logical. Optional. Whether to impute missing values in a way
that will not exclude values from filtering.}

\item{transform}{Logical. Optional. Whether to transform the values to be
consistent with the optimization goal. See \code{\link[=transform.score_vec]{transform.score_vec()}}.}

\item{...}{Currently unused; must be empty.}
}
\value{
A data frame containing importance metrics with one row per term and
one column per metric.
}
\description{
Tidy importance metrics
}
