% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_max_diff.R
\docType{data}
\name{filter_max_diff}
\alias{filter_max_diff}
\title{ANOVA filters}
\format{
An object of class \code{filter_method_max_diff} (inherits from \code{filter_method}) of length 6.
}
\usage{
filter_max_diff
}
\description{
ANOVA filters
}
\keyword{datasets}
